Moved all exercises to central logic because they all shared the same code
